Extract composite signal runtime into echo-signals
