+++ packages/core/echo/echo-schema/src/object/object.ts
@@ -17,7 +17,7 @@ import {
 
 import { type EchoObject, base, debug, db, subscribe } from './types';
 import { type EchoDatabase } from '../database';
-import { compositeRuntime } from '../util/signal';
+import { compositeRuntime } from '@dxos/echo-signals/runtime';
 
 /**
  * Base class for all echo objects.
