+++ packages/core/echo/echo-schema/src/util/index.ts
@@ -5,4 +5,3 @@
 export * from './accessor';
 export * from './devtools-formatter';
 export * from './refs';
-export * from './signal';
