+++ packages/core/echo/echo-schema/tsconfig.json
@@ -59,6 +59,9 @@
       "path": "../echo-pipeline"
     },
     {
+      "path": "../echo-signals"
+    },
+    {
       "path": "../echo-typegen"
     },
     {
