+++ packages/core/echo/echo-signals/package.json
@@ -25,6 +25,12 @@
       "import": "./dist/lib/browser/react.mjs",
       "require": "./dist/lib/node/react.cjs",
       "node": "./dist/lib/node/react.cjs"
+    },
+    "./runtime": {
+      "browser": "./dist/lib/browser/runtime/index.mjs",
+      "import": "./dist/lib/browser/runtime/index.mjs",
+      "require": "./dist/lib/node/runtime/index.cjs",
+      "node": "./dist/lib/node/runtime/index.cjs"
     }
   },
   "types": "dist/types/src/index.d.ts",
@@ -35,6 +41,9 @@
       ],
       "react": [
         "dist/types/src/react.d.ts"
+      ],
+      "runtime": [
+        "dist/types/src/runtime/index.d.ts"
       ]
     }
   },
@@ -43,7 +52,6 @@
     "src"
   ],
   "dependencies": {
-    "@dxos/echo-schema": "workspace:*",
     "@preact/signals": "^1.2.1",
     "@preact/signals-core": "^1.5.0",
     "@preact/signals-react": "^1.3.6"
