+++ packages/core/echo/echo-signals/project.json
@@ -8,7 +8,8 @@
         "entryPoints": [
           "packages/core/echo/echo-signals/src/core.ts",
           "packages/core/echo/echo-signals/src/index.ts",
-          "packages/core/echo/echo-signals/src/react.ts"
+          "packages/core/echo/echo-signals/src/react.ts",
+          "packages/core/echo/echo-signals/src/runtime/index.ts"
         ]
       }
     },
