+++ packages/core/echo/echo-signals/src/preact.ts
@@ -4,7 +4,7 @@
 
 import { signal, batch } from '@preact/signals';
 
-import { registerSignalRuntime as registerRuntimeForEcho } from '@dxos/echo-schema';
+import { registerSignalRuntime as registerRuntimeForEcho } from './runtime';
 
 const registered = false;
