+++ packages/core/echo/echo-signals/tsconfig.json
@@ -12,9 +12,5 @@
   "include": [
     "src"
   ],
-  "references": [
-    {
-      "path": "../echo-schema"
-    }
-  ]
+  "references": []
 }
